QuQCS/pyproject.toml
[tool.poetry]
name = "quqcs"
version = "0.1.0"
description = "quqcs is an open source library for quantum compute simulating on NVIDIA GPU"
authors = ["sunhaiyang <sunhaiyang@qudoor.cn>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "http://www.qudoor.com/"
keywords = ["cuQuantum", "quantum"]
exclude = ["GPU_cuQuantum/test"]

classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: Apache Software License"
]

packages = [
    { include = "qucuQuantum" }
]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
numpy = "^1.23.5"
cuquantum = "^22.11.0"
cuquantum-python = "^22.11.0"
cupy-cuda11x = "^11.4.0"
cutensornet-cu11 = "^2.0.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"